requirements.txt
numpy
faiss-cpu
PyPDF2
sentence-transformers
google-generativeai
